requirements.txt
rumps>=0.4.0
Pillow>=10.0.0
speedtest-cli>=2.1.3
pyobjc-framework-CoreWLAN>=10.0
pyobjc-framework-CoreLocation>=10.0
tplinkrouterc6u>=4.0.0
